deployment initial commit

+++ backend/controllers/passwordController.js
@@ -1,5 +1,6 @@
 import { getCollection } from "../config/db.js";
 import { encryptData, decryptData } from "../utils/crypto.js";
+import { validatePasswordPayload } from "../utils/validation.js";
 
 export async function getPasswords(req,res){
     const collection = getCollection("passwords");
@@ -16,6 +17,10 @@ export async function getPasswords(req,res){
 export async function addPassword(req,res){
     const collection = getCollection("passwords");
     const password = req.body;
+    const validationError = validatePasswordPayload(password);
+    if (validationError) {
+        return res.status(400).json({ message: validationError });
+    }
 
     password.userId = req.userId;
 
@@ -34,6 +39,9 @@ export async function deletePassword(req,res){
     console.log(collection)
 
     const id = req.params.id;
+    if (!id) {
+        return res.status(400).json({ message: "Password id is required" });
+    }
     console.log(id)
 
     const result = await collection.deleteOne({
@@ -51,8 +59,15 @@ export async function updatePassword(req,res){
     const collection = getCollection("passwords");
 
     const id = req.params.id;
+    if (!id) {
+        return res.status(400).json({ message: "Password id is required" });
+    }
 
     const updatedPassword = req.body;
+    const validationError = validatePasswordPayload(updatedPassword);
+    if (validationError) {
+        return res.status(400).json({ message: validationError });
+    }
     
     updatedPassword.password = encryptData(updatedPassword.password);
     console.log(updatedPassword)
